feat: "닉네임 변경 코드 작성"

+++ src/components/common/login/LoginButton.tsx
@@ -1,6 +1,7 @@
 import { GoogleAuthProvider, signInWithPopup, FacebookAuthProvider } from "firebase/auth";
+import { doc, getDoc, setDoc } from 'firebase/firestore';
 import { useNavigate } from "react-router-dom";
-import { firebaseAuth } from "../../../../firebase";
+import { firebaseAuth, firebaseDb } from "../../../../firebase";
 import { useUserStore } from "../../../stores/store";
 
 export default function LoginButton() {
@@ -11,35 +12,76 @@ export default function LoginButton() {
   const handleGoogleLogin = async () => {
     try {
       const result = await signInWithPopup(firebaseAuth, googleProvider);
+      const uid = result.user.uid;
       const displayName = result.user.displayName ?? 'user';
-      const uid = result.user.uid
+      const photoURL = result.user.photoURL ?? 'user';
 
-      setUser(displayName, uid);
+      const userRef = doc(firebaseDb, 'userData', uid);
+      const docSnap = await getDoc(userRef);
+      // 최초로그인
+      if (!docSnap.exists()) {
+        await setDoc(userRef, {
+          name: displayName,
+          photo: photoURL,
+        });
+
+        setUser(displayName, uid, photoURL);
+
+        sessionStorage.setItem('uid', uid);
+        sessionStorage.setItem('name', displayName);
+        sessionStorage.setItem('photo', photoURL);
+      } else {
+        // 재로그인
+        const userData = docSnap.data();
+        setUser(userData.name, uid, userData.photo);
+        sessionStorage.setItem('uid', uid);
+        sessionStorage.setItem('name', userData.name);
+        sessionStorage.setItem('photo', userData.photo);
+      }
+
       navigate('/');
-
-      sessionStorage.setItem('uid', uid);
-      sessionStorage.setItem('name', displayName);
-
-      console.log("Logged in as:", displayName, uid);
     } catch (error) {
-      console.error("Error during Google login:", error);
+      console.error("Google 로그인 중 오류:", error);
     }
   };
 
   const handleFacebookLogin = async () => {
+    // 페이스북 로그인 처리 (구글 로그인과 비슷한 구조)
     try {
       const result = await signInWithPopup(firebaseAuth, new FacebookAuthProvider());
-      const uid = result.user.uid
+      const uid = result.user.uid;
       const displayName = result.user.displayName ?? 'user';
+      const photoURL = result.user.photoURL ?? 'user';
 
-      sessionStorage.setItem('uid', uid);
-      sessionStorage.setItem('name', displayName);
+      const userRef = doc(firebaseDb, 'userData', uid);
+      const docSnap = await getDoc(userRef);
+      // 최초로그인
+      if (!docSnap.exists()) {
+        await setDoc(userRef, {
+          name: displayName,
+          photo: photoURL,
+        });
+
+        setUser(displayName, uid, photoURL);
+
+        sessionStorage.setItem('uid', uid);
+        sessionStorage.setItem('name', displayName);
+        sessionStorage.setItem('photo', photoURL);
+      } else {
+        // 재로그인
+        const userData = docSnap.data();
+        setUser(userData.name, uid, userData.photo);
+        sessionStorage.setItem('uid', uid);
+        sessionStorage.setItem('name', userData.name);
+        sessionStorage.setItem('photo', userData.photo);
+      }
 
       navigate('/');
     } catch (error) {
-      console.error("Error during Facebook login:", error);
+      console.error("Google 로그인 중 오류:", error);
     }
   };
+
 
   return (
     <div className="flex flex-col gap-[8px] w-full item-center">
@@ -58,4 +100,3 @@ export default function LoginButton() {
     </div>
   );
 }
-
